% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert-is-on-os-path.R, R/is-on-os-path.R
\name{assert_all_are_on_os_path}
\alias{assert_all_are_on_os_path}
\alias{assert_any_are_on_os_path}
\alias{is_on_os_path}
\title{Is the path on the OS path?}
\usage{
assert_all_are_on_os_path(
  x,
  severity = getOption("assertive.severity", "stop")
)

assert_any_are_on_os_path(
  x,
  severity = getOption("assertive.severity", "stop")
)

is_on_os_path(x, .xname = get_name_in_parent(x))
}
\arguments{
\item{x}{An path to check.}

\item{severity}{How severe should the consequences of the assertion be?
Either \code{"stop"}, \code{"warning"}, \code{"message"}, or \code{"none"}.}

\item{.xname}{Not intended to be used directly.}
}
\value{
\code{TRUE} if the specified paths are on the OS search path.
}
\description{
Is the specified path on the operating system search path?
}
\note{
The OS search path is determined with \code{Sys.getenv("path")}.  For
files, the path of the containing folder is checked rather than the path of 
the file itself.
}
\examples{
is_on_os_path(
  c(R.home("bin"), R.home("etc"), "a nonexistent path")
) # probably c(TRUE, FALSE, FALSE)
}
