% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-ide.R
\name{is_rstudio_desktop}
\alias{is_rstudio_desktop}
\alias{is_rstudio_server}
\title{Is RStudio running in desktop or server mode?}
\usage{
is_rstudio_desktop()

is_rstudio_server()
}
\description{
Checks for RStudio desktop or server version.
}
\examples{
is_rstudio_desktop()
is_rstudio_server()
}
\references{
The values that RStudio uses for its mode are defined in
\url{https://github.com/rstudio/rstudio/blob/master/src/cpp/session/include/session/SessionConstants.hpp}
via the constants \code{kSessionProgramModeDesktop} and 
\code{kSessionProgramModeServer}.
}
\seealso{
\code{\link{is_rstudio}}, \code{\link{is_rstudio_current}}
}
