% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert-r-has-capability.R, R/r-has-capability.R
\name{assert_r_has_jpeg_capability}
\alias{assert_r_has_jpeg_capability}
\alias{assert_r_has_png_capability}
\alias{assert_r_has_tiff_capability}
\alias{assert_r_has_tcltk_capability}
\alias{assert_r_has_x11_capability}
\alias{assert_r_has_aqua_capability}
\alias{assert_r_has_http_ftp_capability}
\alias{assert_r_has_sockets_capability}
\alias{assert_r_has_libxml_capability}
\alias{assert_r_has_fifo_capability}
\alias{assert_r_has_cledit_capability}
\alias{assert_r_has_iconv_capability}
\alias{assert_r_has_nls_capability}
\alias{assert_r_has_rprof_capability}
\alias{assert_r_has_profmem_capability}
\alias{assert_r_has_cairo_capability}
\alias{assert_r_has_icu_capability}
\alias{assert_r_has_long_double_capability}
\alias{assert_r_has_libcurl_capability}
\alias{r_has_jpeg_capability}
\alias{r_has_capability}
\alias{r_has_png_capability}
\alias{r_has_tiff_capability}
\alias{r_has_tcltk_capability}
\alias{r_has_x11_capability}
\alias{r_has_aqua_capability}
\alias{r_has_http_ftp_capability}
\alias{r_has_sockets_capability}
\alias{r_has_libxml_capability}
\alias{r_has_fifo_capability}
\alias{r_has_cledit_capability}
\alias{r_has_iconv_capability}
\alias{r_has_nls_capability}
\alias{r_has_rprof_capability}
\alias{r_has_profmem_capability}
\alias{r_has_cairo_capability}
\alias{r_has_icu_capability}
\alias{r_has_long_double_capability}
\alias{r_has_libcurl_capability}
\title{Does R have a capability?}
\usage{
assert_r_has_jpeg_capability(
  severity = getOption("assertive.severity", "stop")
)

assert_r_has_png_capability(severity = getOption("assertive.severity", "stop"))

assert_r_has_tiff_capability(
  severity = getOption("assertive.severity", "stop")
)

assert_r_has_tcltk_capability(
  severity = getOption("assertive.severity", "stop")
)

assert_r_has_x11_capability(severity = getOption("assertive.severity", "stop"))

assert_r_has_aqua_capability(
  severity = getOption("assertive.severity", "stop")
)

assert_r_has_http_ftp_capability(
  severity = getOption("assertive.severity", "stop")
)

assert_r_has_sockets_capability(
  severity = getOption("assertive.severity", "stop")
)

assert_r_has_libxml_capability(
  severity = getOption("assertive.severity", "stop")
)

assert_r_has_fifo_capability(
  severity = getOption("assertive.severity", "stop")
)

assert_r_has_cledit_capability(
  severity = getOption("assertive.severity", "stop")
)

assert_r_has_iconv_capability(
  severity = getOption("assertive.severity", "stop")
)

assert_r_has_nls_capability(severity = getOption("assertive.severity", "stop"))

assert_r_has_rprof_capability(
  severity = getOption("assertive.severity", "stop")
)

assert_r_has_profmem_capability(
  severity = getOption("assertive.severity", "stop")
)

assert_r_has_cairo_capability(
  severity = getOption("assertive.severity", "stop")
)

assert_r_has_icu_capability(severity = getOption("assertive.severity", "stop"))

assert_r_has_long_double_capability(
  severity = getOption("assertive.severity", "stop")
)

assert_r_has_libcurl_capability(
  severity = getOption("assertive.severity", "stop")
)

r_has_jpeg_capability()

r_has_png_capability()

r_has_tiff_capability()

r_has_tcltk_capability()

r_has_x11_capability()

r_has_aqua_capability()

r_has_http_ftp_capability()

r_has_sockets_capability()

r_has_libxml_capability()

r_has_fifo_capability()

r_has_cledit_capability()

r_has_iconv_capability()

r_has_nls_capability()

r_has_rprof_capability()

r_has_profmem_capability()

r_has_cairo_capability()

r_has_icu_capability()

r_has_long_double_capability()

r_has_libcurl_capability()
}
\arguments{
\item{severity}{How severe should the consequences of the assertion be?
Either \code{"stop"}, \code{"warning"}, \code{"message"}, or \code{"none"}.}
}
\value{
The \code{is_*} functions return \code{TRUE} if R has the capability 
and \code{FALSE} (with a cause) otherwise.
The \code{assert_*} functions return nothing but throw an error if the 
corresponding \code{is_*} function returns \code{FALSE}.
}
\description{
Check to see if R has a specific capability.
}
\examples{
\dontrun{
if(r_has_png_capability())
{
  png("test.png")
  with(cars, plot(speed, dist))
  dev.off()
} else 
{
  pdf("test.pdf")
  with(cars, plot(speed, dist))
  dev.off()
}
}
}
\seealso{
\code{\link[base]{capabilities}}
}
